config: split off editing into editor.go
